Separate client persistence from client registry notification
This is an attempt to separate the client persistence logic done in
ClientRegistryImpl from notifying the ClientRegistryListeners.

The concerns related to the listeners, are now separated in
two main interfaces: ``ClientRegistryListenerManagement`` (for add/remove
operations) and ``ClientRegistryNotification`` (for the listener
notifications). These can be provided either in LeshanServer or via the
LeshanServerBuilder. The notifications are triggered only by the
RegistrationResource and the client registry is used only in a straight
"persistence" manner in RegistrationHandler.

I also externalized the functionality to allow/disallow
client registrations in a ``RegistrationAcceptor`` interface which can
also be passed to the LeshanServerBuilder and externalized the
RegistrationCleaner in a separate class.

Signed-off-by: Alexander Ellwein <alexander.ellwein@bosch-si.com>