Queue Mode: base interfaces for ongoing implementation
Intro
=====

A "queue mode" is an operation mode of LWM2M protocol (since the early versions
of Lightweight M2M 1.0 Specification), which describes a special operation mode
of a server/client if a client is not always expected to be online.

The specification states that:

* A Client has to register using specified binding mode (e.g. using "UQ" for
  "UDP+Queue");
* A Server has to understand this binding mode (it is *mandatory*);
* A client is not expected to be always online - instead, a server must queue
  the requests and only allowed to send downlink requests one by one to the client
  when server received client's registration update or notification.

This particular pull request contains classes splitted in two main packages:

- **org.eclipse.leshan.server.queue** - contains base interfaces to operate
  with queue mode, i.e. public API for sending downlink requests and managing
  the queue;
- **org.eclipse.leshan.server.queue.persistence** - contains interfaces to
  implement own queue persistence (we will provide only a basic in-memory
  persistence).

org.eclipse.leshan.server.queue - Queue Mode Public API
=======================================================

QueueRequestSender
------------------

``QueueRequestSender`` is an entry point for queue mode related sending.
This is basically an extension of ``LwM2mRequestSender``, which is capable
to maintain a queue of requests for every client registered with queue
binding mode and provides a basic queue management by means of
``QueueManagement`` interface.

QueueManagement
---------------

Queue Management is a public interface for interacting with the queue. It
can be used, e.g. to show QueueRequests for a particular client, drop a
single request or even a whole client's queue.

QueueRequest
------------

A single element of a queue of ``QueueRequestSender`` is represented by a
``QueueRequest``. Amongst the downlink request and client's endpoint, it
contains an information about request's state (``QueueRequestState``),
ID of the request and also an expiration date of the queue request, i.e.
the time point when the queue request should not be considered/send anymore.

Note: this is the user's view on a request, for persistence's view see ``QueueRequestEntity`` below.

QueueRequestState
-----------------

Every ``QueueRequest`` is passed through different states during its life
cycle. ``QueueRequestState`` enumeration describes following states:

1. A fresh created queue request which is sent via ``QueueRequestSender`` is
   put in the queue with the state ``ENQUEUED``. An enqueued request remains in
   this state, waiting to be send.

2. A client's registration update or notify triggers a queue request to be put
   in ``PROCESSING`` state. This indicates that the queue request is currently being
   sent. For every client, there can be only one request in a processing state,
   because all of the requests are processed strictly sequential (per client).

3.
   a) If a ``PROCESSING`` request is timed out while sending, it returns to the
   queue with the state ``DEFERRED``. This means, that after next client's
   update/notify this request will be sent as first (unless the send
   expiration interval is already elapsed, in this case it will be put into ``TTL_ELAPSED``
   state as described in 4)

   b) If a ``PROCESSING`` request is sent, then it is put into ``EXECUTED``
   state. The request will be then put into ``EXECUTED`` state.

   c) If a ``PROCESSING`` request has been deleted by the user while
      sending, it is put into ``PROCESSING_DELETED`` state. That is,
      the request will still be processed until the successful or
      unsucessful result of the sending operation, but will either
      be promoted to ``EXECUTED`` state and then deleted or deleted
      immediately upon a timeout of the send operation.

4. A ``DEFERRED`` queue request, which could not be sent within its expiration
   time, will be put in ``TTL_ELAPSED`` state.

5. Queue requests which are in states ``EXECUTED`` or ``TTL_ELAPSED`` may be
   removed from the queue (or archived for some period of time, depending on
   the queue's concrete implementation).

QueueRequestDeletedException
----------------------------

A ``QueueRequestDeletedException`` is needed to notify the sender on an explicit
removal of a ``QueueRequest``. It is thrown in the ``ErrorCallback`` registered
in the ``QueueRequestSender``.

org.eclipse.leshan.server.queue.persistence - Persistence Interfaces
====================================================================

QueuePersistence
----------------

``QueuePersistence`` provides basic operations for adding, retrieving, updating
and deleting ``QueueRequestEntity`` elements.

QueueRequestEntity
------------------

``QueueRequestEntity`` represents a persistence object for ``QueueRequest``.

QueueRequestEntityFactory
-------------------------

An interface for a factory, able to create ``QueueRequestEntities``, based
on given client's endpoint, downlink request and given Time-To-Live for the
request.

ExpirationChecker
-----------------

An ``ExpirationChecker`` allows to check an expiration of a particular
queue request and calculates the absolute time stamp for a queue request
expiration, note that the implementation of ``ExpirationChecker`` may be
dependent or specific to the ``QueuePersistence`` implementation.

Also-by: Daniel Maier <daniel.maier@bosch-si.com>

Signed-off-by: Alexander Ellwein <alexander.ellwein@bosch-si.com>