Check BS authentication before computing BS config
Currently, BS configuration to send to a device is
computed before checking that the BS credentials
sent by the device are valid.

When rotating BS credentials, a server might generate
new credentials as a result of checking credentials.

Changing the order or the calls makes it possible
to send newly generated BS credentials as soon as
they are generated.

Signed-off-by: Pierre-Henri Trivier <phtrivier@sierrawireless.com>