Use Endpoint's URI instead of socket address for Client registration.
The URI includes the scheme of the protocol the client used for
registering with the LWM2M server. Before this change, this information
had to be deduced by comparing the port of the registration address with
the port of the "secure endpoint address" that has to be kept on record
for that purpose.