Fix #463 : Remove optimization for encoding single instance object.
This optimization was removed from the specification.
The spec say now :
For simplicity and to prevent any ambiguity on Object Instance Identity,
when the Object Instance ID is not specified in the request, the Object
Instance TLV MUST be used, whatever the number of Object Instances (1 or
more) to return to the LwM2M Server.