Implements an identity and matching of the identity for OSCORE
An identity string based on OSCORE parameters is now set for
connected clients that use OSCORE. The identity string contains
the OSCORE Sender and Recipient ID the client and server are using.
These parameters can be retrieved from the source EndpointContext.

The identity for OSCORE is also set as a SecurityInfo based on the
OSCORE context configured on the server to use with a specific client.

The identity of a connected client and the configured SecurityInfo are
then matched in the SecurityChecker.

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>