#726: Generate OSCORE object and use it for endpoint creation
Code to implement the points listed here:
https://github.com/eclipse/leshan/pull/718#issuecomment-521316754

The OSCORE object is generated and filled in the Leshan client demo.
It will then be used to create a ServerInfo object using the
ServersInfoExtractor. Next the ServerInfo is used in the
CaliforniumEndpointsManager to generate endpoints towards servers
that OSCORE should be used with.

See also issue regarding OSCORE client integration:
https://github.com/eclipse/leshan/issues/726

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>