Add `shutdownTrigger` for registration engine
If the procedure inside of `ClientInitiatedBootstrapTask`, `UpdateRegistrationTask`,
and/or `RegistrationTask` raises an unexpected exception, that exception caught by
a `catch` clause and it logs that, but that task never works after caught.
For example, if `registerWithRetry()` that is in `RegistrationTask#run()` raises a
`RuntimeException`, that task never scheduled anymore.
This behavior makes a little bit hard to implement a robust LwM2M client.
So this patch introduces the `shutdownTrigger` mechanism to shut down the client when
an unexpected exception occurred. `shutdownTrigger` is an implementation of `Runnable`,
so the user can hand over the user-defined shutdown procedure to the registration engine.

Signed-off-by: moznion <moznion@gmail.com>